pyproject.toml
[project]
name = "voyager"
version = "0.1.0"
description = "Meta-skills for Coding Agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.18",
    "jsonschema>=4.25.1",
    "pyyaml>=6.0",
    "typer>=0.15.0",
]

[project.optional-dependencies]
retrieval = [
    "ragatouille>=0.0.8",
]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
]

[project.scripts]
voyager = "voyager.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voyager"]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
